server/agent/requirements.txt
google-adk>=1.31.0
google-genai>=1.0.0
httpx>=0.27.0
tenacity>=8.2.0
python-docx>=1.1.0
openpyxl>=3.1.0
